introduce WorkspaceTranspilationTrace
