+++ playground/src/main/kotlin/com/strumenta/kolasu/playground/WorkspaceTranspilationTrace.kt
@@ -1,0 +1,17 @@
+package com.strumenta.kolasu.playground
+
+import com.strumenta.kolasu.model.Node
+import com.strumenta.kolasu.validation.Issue
+import com.strumenta.kolasu.validation.Result
+
+/**
+ * A transpilation trace can be visualized to demonstrate how the transpiler work.
+ * This transpilation trace represent a set of files being transformed in a second set of files.
+ */
+class WorkspaceTranspilationTrace {
+    data class WorkspaceFile(val path: String, val code: String, val result: Result<Node>)
+
+    val originalFiles : List<WorkspaceFile> = mutableListOf()
+    val generatedFiles : List<WorkspaceFile> = mutableListOf()
+    val transpilationIssues: List<Issue> = emptyList()
+}
